requirements.txt
colorthief==0.2.1
Pillow==7.1.2
opencv-python==4.2.0.32
mss==6.1.0
numpy==1.18.1
pyautogui==0.9.48